+++ .gitignore
@@ -1,0 +1,2 @@
+RCA_YOLOv8.egg-info/
+dist/
